Refactor order test cases
